EmbeddedClientFactory now creates embedded database with isPersitent=false by default.
When operating in non-persistent mode, the file system is still checked for stores, thus
giving the impression that we're able to operate in read-only mode, which is not really
the case, as while the Store objects will be loaded, there will be no pager able to
load the persisted indexes and pages. (this should be tidied up soon).

This can be overridden, and is used just for PersistentTest for now.  This is done by
calling EmbeddedClientFactory.setPesistent(true) before calling getInstance().