Merge NonPersistentAttrDefinitionMgr into AttrDefinitionMgr and allow readOnly mode
NonPersistentAttrDefinitionMgr was misleading, as it was not by definition non-persistent
and could be used in either way.

ADM now has public ctor, and can also be set to prevent modification, so that it can
be used in a scenario where the configuration should not be changed during the
running of the application.