Add m2e >= 0.13 interim config for working nicely with AspectJ modules
m2e AJDT integration is only working up to 0.12 release.  This commit can be reverted
when it is available.

Some aspects (sorry) of importing a maven project won't be as expected until
this is provided, in that the m2e configurator doesn't know to add
the src/main/aspect folder to the eclipse build path.  This will have
to be added manually.