Add RandomAttributeSource to ease prototyping
This gives access to the random value generators that were
recently added to the attrs module, and is intended to support
the generation of data for building large data sets for testing.

The intention here is to make it very easy for someone to take
a sample app based, on say a dating scenario, and quickly
be able to adapt that to a different service.