Issue #62: Add assertValid() to all Scorers (matchers)
Invalid scorers were previously not discovered until an attempt was
made to use them.  Now they are validated as score configurations
are updated.

Far nicer than receiving an exception from the server when doing
a search :)

Also added missing weight attribute to one scorer that this
showed to be mis-configured.