Add FuzzyProperty.isFuzzyAttribute()
This solves single enum problem and takes a first step to moving attrDef stuff
into the mapping context so we can use attrId from the def.

If a String has been defined as an enum attribute in xml then it is treated as
enum, otherwise just as dumb non-indexed String.